service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
httpx==0.28.1
pillow==11.3.0
google-generativeai==0.8.5
